+++ context_get_headers.go
@@ -1,0 +1,9 @@
+// +build go1.14
+
+package gear
+
+import "net/http"
+
+func getHeaderValues(h http.Header, key string) []string {
+	return h.Values(key)
+}