+++ context_get_headers_go1_13.go
@@ -1,0 +1,12 @@
+// +build !go1.14
+
+package gear
+
+import (
+	"net/http"
+	"net/textproto"
+)
+
+func getHeaderValues(h http.Header, key string) []string {
+	return h[textproto.CanonicalMIMEHeaderKey(key)]
+}